+++ src/main/java/frc/robot/core/Robot.java
@@ -3,6 +3,7 @@ package frc.robot.core;
 
 import java.util.ArrayList;
 
+import edu.wpi.first.cameraserver.CameraServer;
 import frc.robot.utils.FieldCentric;
 import frc.robot.utils.PressureRegulator;
 import org.montclairrobotics.sprocket.SprocketRobot;
@@ -23,6 +24,7 @@ import org.montclairrobotics.sprocket.motors.Module;
 import org.montclairrobotics.sprocket.motors.Motor;
 import org.montclairrobotics.sprocket.motors.SEncoder;
 import org.montclairrobotics.sprocket.pipeline.Step;
+import org.montclairrobotics.sprocket.utils.CameraServers;
 import org.montclairrobotics.sprocket.utils.Debug;
 import org.montclairrobotics.sprocket.utils.PID;
 
@@ -111,6 +113,7 @@ public class Robot extends SprocketRobot {
         visionCorrection.setTarget(200); // TODO: Test and tune
         new ToggleButton(Control.driveStick, Control.Port.AUTO_HATCH, visionCorrection);
         steps.add(visionCorrection);
+        steps.add(new VisionCorrection(new VisionTarget(CameraServer.getInstance().getVideo()), new PID(.1, 0, 0)));
         steps.add(correction);
         steps.add(fieldCentric);
         steps.add(orientation);
