+++ src/main/java/frc/robot/utils/VisionTarget.java
@@ -1,0 +1,197 @@
+package frc.robot.utils;
+import java.io.File;
+import java.io.FileWriter;
+import java.io.IOException;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Map;
+import java.util.stream.Collectors;
+import java.util.HashMap;
+
+import edu.wpi.cscore.CvSink;
+import org.montclairrobotics.sprocket.geometry.Vector;
+import org.montclairrobotics.sprocket.geometry.XY;
+import org.montclairrobotics.sprocket.loop.Updatable;
+import org.montclairrobotics.sprocket.utils.Input;
+import org.opencv.core.*;
+import org.opencv.core.Core.*;
+import org.opencv.features2d.FeatureDetector;
+import org.opencv.imgcodecs.Imgcodecs;
+import org.opencv.imgproc.*;
+
+public class VisionTarget implements Input<Double>, Updatable {
+    //Outputs
+    private Mat hslThresholdOutput = new Mat();
+    private ArrayList<MatOfPoint> findContoursOutput = new ArrayList<MatOfPoint>();
+    private ArrayList<MatOfPoint> filterContoursOutput = new ArrayList<MatOfPoint>();
+    private Vector out;
+
+    private CvSink source;
+
+    static {
+        System.loadLibrary(Core.NATIVE_LIBRARY_NAME);
+    }
+
+    public VisionTarget(CvSink source){
+        this.source = source;
+    }
+
+    /**
+     * This is the primary method that runs the entire pipeline and updates the outputs.
+     */
+    public void process(Mat source0) {
+        // Step HSL_Threshold0:
+        Mat hslThresholdInput = source0;
+        double[] hslThresholdHue = {0.0, 144.36860068259384};
+        double[] hslThresholdSaturation = {0.0, 157.0904436860068};
+        double[] hslThresholdLuminance = {220.14388489208633, 255.0};
+        hslThreshold(hslThresholdInput, hslThresholdHue, hslThresholdSaturation, hslThresholdLuminance, hslThresholdOutput);
+
+        // Step Find_Contours0:
+        Mat findContoursInput = hslThresholdOutput;
+        boolean findContoursExternalOnly = false;
+        findContours(findContoursInput, findContoursExternalOnly, findContoursOutput);
+
+        // Step Filter_Contours0:
+        ArrayList<MatOfPoint> filterContoursContours = findContoursOutput;
+        double filterContoursMinArea = 1000.0;
+        double filterContoursMinPerimeter = 100.0;
+        double filterContoursMinWidth = 0.0;
+        double filterContoursMaxWidth = 1000.0;
+        double filterContoursMinHeight = 0.0;
+        double filterContoursMaxHeight = 1000.0;
+        double[] filterContoursSolidity = {69.24460431654677, 100};
+        double filterContoursMaxVertices = 1000000.0;
+        double filterContoursMinVertices = 0.0;
+        double filterContoursMinRatio = 0.0;
+        double filterContoursMaxRatio = 1000.0;
+        filterContours(filterContoursContours, filterContoursMinArea, filterContoursMinPerimeter, filterContoursMinWidth, filterContoursMaxWidth, filterContoursMinHeight, filterContoursMaxHeight, filterContoursSolidity, filterContoursMaxVertices, filterContoursMinVertices, filterContoursMinRatio, filterContoursMaxRatio, filterContoursOutput);
+
+    }
+
+    /**
+     * This method is a generated getter for the output of a HSL_Threshold.
+     * @return Mat output from HSL_Threshold.
+     */
+    public Mat hslThresholdOutput() {
+        return hslThresholdOutput;
+    }
+
+    /**
+     * This method is a generated getter for the output of a Find_Contours.
+     * @return ArrayList<MatOfPoint> output from Find_Contours.
+     */
+    public ArrayList<MatOfPoint> findContoursOutput() {
+        return findContoursOutput;
+    }
+
+    /**
+     * This method is a generated getter for the output of a Filter_Contours.
+     * @return ArrayList<MatOfPoint> output from Filter_Contours.
+     */
+    public ArrayList<MatOfPoint> filterContoursOutput() {
+        return filterContoursOutput;
+    }
+
+
+    /**
+     * Segment an image based on hue, saturation, and luminance ranges.
+     *
+     * @param input The image on which to perform the HSL threshold.
+     * @param hue The min and max hue
+     * @param sat The min and max saturation
+     * @param lum The min and max luminance
+     */
+    private void hslThreshold(Mat input, double[] hue, double[] sat, double[] lum,
+                              Mat out) {
+        Imgproc.cvtColor(input, out, Imgproc.COLOR_BGR2HLS);
+        Core.inRange(out, new Scalar(hue[0], lum[0], sat[0]),
+                new Scalar(hue[1], lum[1], sat[1]), out);
+    }
+
+    /**
+     * Sets the values of pixels in a binary image to their distance to the nearest black pixel.
+     * @param input The image on which to perform the Distance Transform.
+     */
+    private void findContours(Mat input, boolean externalOnly,
+                              List<MatOfPoint> contours) {
+        Mat hierarchy = new Mat();
+        contours.clear();
+        int mode;
+        if (externalOnly) {
+            mode = Imgproc.RETR_EXTERNAL;
+        }
+        else {
+            mode = Imgproc.RETR_LIST;
+        }
+        int method = Imgproc.CHAIN_APPROX_SIMPLE;
+        Imgproc.findContours(input, contours, hierarchy, mode, method);
+    }
+
+
+    /**
+     * Filters out contours that do not meet certain criteria.
+     * @param inputContours is the input list of contours
+     * @param output is the the output list of contours
+     * @param minArea is the minimum area of a contour that will be kept
+     * @param minPerimeter is the minimum perimeter of a contour that will be kept
+     * @param minWidth minimum width of a contour
+     * @param maxWidth maximum width
+     * @param minHeight minimum height
+     * @param maxHeight maximimum height
+     * @param minVertexCount minimum vertex Count of the contours
+     * @param maxVertexCount maximum vertex Count
+     * @param minRatio minimum ratio of width to height
+     * @param maxRatio maximum ratio of width to height
+     */
+    private void filterContours(List<MatOfPoint> inputContours, double minArea,
+                                double minPerimeter, double minWidth, double maxWidth, double minHeight, double
+                                        maxHeight, double[] solidity, double maxVertexCount, double minVertexCount, double
+                                        minRatio, double maxRatio, List<MatOfPoint> output) {
+        final MatOfInt hull = new MatOfInt();
+        output.clear();
+        //operation
+        for (int i = 0; i < inputContours.size(); i++) {
+            final MatOfPoint contour = inputContours.get(i);
+            final Rect bb = Imgproc.boundingRect(contour);
+            if (bb.width < minWidth || bb.width > maxWidth) continue;
+            if (bb.height < minHeight || bb.height > maxHeight) continue;
+            final double area = Imgproc.contourArea(contour);
+            if (area < minArea) continue;
+            if (Imgproc.arcLength(new MatOfPoint2f(contour.toArray()), true) < minPerimeter) continue;
+            Imgproc.convexHull(contour, hull);
+            MatOfPoint mopHull = new MatOfPoint();
+            mopHull.create((int) hull.size().height, 1, CvType.CV_32SC2);
+            for (int j = 0; j < hull.size().height; j++) {
+                int index = (int)hull.get(j, 0)[0];
+                double[] point = new double[] { contour.get(index, 0)[0], contour.get(index, 0)[1]};
+                mopHull.put(j, 0, point);
+            }
+            final double solid = 100 * area / Imgproc.contourArea(mopHull);
+            if (solid < solidity[0] || solid > solidity[1]) continue;
+            if (contour.rows() < minVertexCount || contour.rows() > maxVertexCount)	continue;
+            final double ratio = bb.width / (double)bb.height;
+            if (ratio < minRatio || ratio > maxRatio) continue;
+            output.add(contour);
+        }
+    }
+
+    @Override
+    public void update() {
+        Mat output = new Mat();
+        source.grabFrame(output);
+        process(output);
+        if(!filterContoursOutput.isEmpty()){
+            Rect r1 = Imgproc.boundingRect(filterContoursOutput.get(0));
+            Rect r2 = Imgproc.boundingRect(filterContoursOutput.get(1));
+            double x = (r1.x + r2.x)/2.0;
+            double y = (r1.y + r2.y)/2.0;
+            out = new XY(x, y);
+        }
+    }
+
+    @Override
+    public Double get() {
+        return out.getX();
+    }
+}
